pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "domainchecker"
version = "1.0.0"
description = "Comprehensive domain checking, WHOIS lookup, and DNS resolution library"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "TaxLien Team", email = "team@taxlien.online"}
]
maintainers = [
    {name = "TaxLien Team", email = "team@taxlien.online"}
]
keywords = ["domain", "whois", "dns", "expiration", "checker"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Networking",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "python-whois>=0.8.0",
    "dnspython>=2.4.0",
    "python-dateutil>=2.9.0",
    "colorama>=0.4.6",
    "concurrent-futures>=3.1.1; python_version<'3.2'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/taxlien-online/domainchecker"
Documentation = "https://domainchecker.readthedocs.io"
Repository = "https://github.com/taxlien-online/domainchecker.git"
"Bug Tracker" = "https://github.com/taxlien-online/domainchecker/issues"

[project.scripts]
domainchecker = "domainchecker.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
